feat(incidentscsv): made functions more abstract
made getcsv and downloadlink functions

re #2292
